backend/pyproject.toml
[project]
name = "digidice"
version = "0.0.1"
description = "Fair Digital Dice"
authors = [
    {name = "Ioannis Somos", email = "i.somos@aueb.gr"},
    {name = "Thomas Katsantas", email = "th.katsantas@aueb.gr"}
]
keywords = ["dice", "game"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django :: 4.2",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.8"
dependencies = [
    "django[argon2]~=4.2",
    "djangorestframework~=3.14",
    "djangorestframework-simplejwt[crypto]~=5.2",
    "mysqlclient~=2.1"
]

[project.optional-dependencies]
uwsgi = ["uwsgi~=2.0"]
local = [
    "django-extensions~=3.2",
    "pyopenssl~=23.2",
    "werkzeug~=2.3"
]
dev = [
    "django-stubs~=4.2",
    "djangorestframework-stubs~=3.14"
]

[project.urls]
Homepage = "https://gitlab.com/ObserverOfTime/netsec-assignment"

[build-system]
requires = ["setuptools>=58.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["keys"]

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]

[tool.django-stubs]
django_settings_module = "digidice.settings"
